requirements.txt
# FastAPI and ASGI server
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.29.0,<1.0.0
python-multipart>=0.0.6,<1.0.0

# TTS Service
deepgram-sdk>=3.3.3,<4.0.0

# Google Cast Integration
pychromecast>=14.0.9,<15.0.0
zeroconf>=0.132.2,<1.0.0

# HTTP Client
httpx>=0.27.0,<1.0.0

# Configuration & Environment
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.7.1,<3.0.0
pydantic-settings>=2.2.1,<3.0.0

# Logging and monitoring
structlog>=24.1.0,<25.0.0
rich>=14.1.0,<15.0.0

# System utilities
psutil>=5.9.0,<6.0.0
aiofiles>=23.0.0,<24.0.0


# Rate Limiting
slowapi>=0.1.9,<1.0.0

# Development & Testing
pytest>=8.2.0,<9.0.0
pytest-asyncio>=0.23.7,<1.0.0
pytest-cov>=5.0.0,<6.0.0
pytest-mock>=3.14.0,<4.0.0
black>=24.4.2,<25.0.0
flake8>=7.0.0,<8.0.0
mypy>=1.9.0,<2.0.0
